Upgrade everything

+++ start.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -eu
+
 # produce all values of environment
 # variables with a name starting with $1
 all_environment_values() {
@@ -11,7 +13,9 @@ all_environment_values() {
 echo Generating config...
 
 # store database connection parameters for pg_dump
-rm ~/.pgpass
+if [ -e ~/.pgpass ]; then
+    rm ~/.pgpass
+fi
 for db in $(all_environment_values DATABASE); do
 	echo $db >> ~/.pgpass
 done
@@ -25,8 +29,10 @@ ssh-keyscan -p $SFTP_PORT $SFTP_HOST > ~/.ssh/known_hosts
 # store backup url
 echo BACKUP_URL=sftp://$SFTP_USER:$SFTP_PASSWORD@$SFTP_HOST:$SFTP_PORT/$BACKUP_NAME > /etc/backup
 
-mkfifo /opt/fifo
-# tigger 'tail -f' to open fifo
+if ! [ -f /opt/fifo ]; then
+	mkfifo /opt/fifo
+fi
+# trigger 'tail -f' to open fifo
 echo Logging started... > /opt/fifo &
 
 echo "00 0 * * * root /opt/backup.sh > /opt/fifo 2>&1" > /etc/crontab
